+++ Spring-JMS/src/main/java/com/ibm/mq/samples/jms/spring/level112/MessageConsumer112.java
@@ -14,13 +14,12 @@
  * limitations under the License.
  */
 
-package com.ibm.mq.samples.jms.spring.Level112;
+package com.ibm.mq.samples.jms.spring.level112;
 
 import com.ibm.mq.samples.jms.spring.globals.data.OurData;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.springframework.jms.annotation.JmsListener;
-import org.springframework.stereotype.Component;
 
 //@Component
 public class MessageConsumer112 {
